BISERVER-6820 - Metadata models generated by AgileBI are broken in PME
Changed how annotations are stored in the xmi file. they are now set as CWM:TaggedValue that are prefixed with "ANNOTATION_". This was to fix the issue where the original way we stored the annotations wasn't supported in the CWMOLAP schema and would cause parse failures when trying to open them in PME.

The caveat to this is that any model (xmi) saved prior to this change will not be support annotations the 'old' way. (no released software supported annotations, so only pre released 4.5 builds would be affected. this was added to support geography roles which is a new feature in 4.5)